+++ packages/css-color-parser/src/functions/lab-normalize-channel-values.ts
@@ -15,7 +15,7 @@ export function normalize_Lab_ChannelValues(token: CSSToken, index: number, colo
 			token[2],
 			token[3],
 			{
-				value: NaN,
+				value: Number.NaN,
 				type: NumberType.Number,
 			},
 		];